requirements.txt
tensorflow==2.16.1
opencv-python==4.8.1.78
mediapipe==0.10.21
numpy==1.26.0
matplotlib==3.8.0
scikit-learn==1.3.2
Pillow==10.1.0